+++ app/lib/l10n/gen/ouds_flutter_app_localizations.dart
@@ -755,12 +755,6 @@ abstract class AppLocalizations {
   /// **'Error'**
   String get app_components_badge_negative_label_a11y;
 
-  /// No description provided for @app_components_badge_disable_label_a11y.
-  ///
-  /// In en, this message translates to:
-  /// **'Not available'**
-  String get app_components_badge_disable_label_a11y;
-
   /// No description provided for @app_components_chip_label.
   ///
   /// In en, this message translates to:
@@ -868,6 +862,12 @@ abstract class AppLocalizations {
   /// In en, this message translates to:
   /// **'Trailing icon content description'**
   String get app_components_textInput_trailingIcon_a11y;
+
+  /// No description provided for @app_components_text_input_helperLinkText_label.
+  ///
+  /// In en, this message translates to:
+  /// **'Helper link'**
+  String get app_components_text_input_helperLinkText_label;
 
   /// No description provided for @app_components_link_label.
   ///