+++ ouds_core/lib/components/badge/ouds_badge.dart
@@ -57,6 +57,7 @@ enum Type {
 ///   status: OudsBadgeStatus.negative,
 ///   size: OudsBadgeSize.large,
 ///   label: '120',
+///   enabled: true,
 ///   icon: 'assets/ic_heart_badge.svg',
 ///   child: Icon(Icons.favorite), // Replace with your child widget";
 /// );
@@ -69,6 +70,7 @@ class OudsBadge extends StatefulWidget {
   final String? label;
   final String? icon;
   final Widget? child;
+  final bool enabled;
   final String? semanticsLabel;
 
   const OudsBadge({
@@ -78,6 +80,7 @@ class OudsBadge extends StatefulWidget {
     this.label,
     this.icon,
     this.child,
+    this.enabled = true,
     this.semanticsLabel
   });
 
@@ -118,13 +121,14 @@ class _OudsBadgeState extends State<OudsBadge> {
       ),
       child: Semantics(
         label: widget.semanticsLabel,
+        enabled: widget.enabled,
         child: Badge(
           padding: widget.icon != null
               ? EdgeInsets.only(left: badge.spaceInset, right: badge.spaceInset)
               : widget.size == OudsBadgeSize.large
                   ? EdgeInsets.only(left: badge.spacePaddingInlineLarge, right: badge.spacePaddingInlineLarge)
                   : EdgeInsets.only(left: badge.spacePaddingInlineMedium, right: badge.spacePaddingInlineMedium),
-          backgroundColor: badgeStatusModifier.getStatusColor(widget.status),
+          backgroundColor: badgeStatusModifier.getStatusColor(widget.status,widget.enabled),
           label: badgeLabel,
           child: widget.child,
         ),
@@ -142,8 +146,8 @@ class _OudsBadgeState extends State<OudsBadge> {
           child: Text(
               _formattedLabel(),
               style: widget.size == OudsBadgeSize.large
-                  ? theme.typographyTokens.typeLabelDefaultMedium(context).copyWith(color: badgeStatusModifier.getStatusTextAndIconColor((widget.status)))
-                  : theme.typographyTokens.typeLabelDefaultSmall(context).copyWith(color: badgeStatusModifier.getStatusTextAndIconColor((widget.status))),
+                  ? theme.typographyTokens.typeLabelDefaultMedium(context).copyWith(color: badgeStatusModifier.getStatusTextAndIconColor((widget.status),widget.enabled))
+                  : theme.typographyTokens.typeLabelDefaultSmall(context).copyWith(color: badgeStatusModifier.getStatusTextAndIconColor((widget.status),widget.enabled)),
               textAlign: TextAlign.center,
             ),
         )
@@ -157,15 +161,18 @@ class _OudsBadgeState extends State<OudsBadge> {
     if (assetName == null) {
       return SizedBox.shrink(); // widget empty
     }
+
+   final icon = badgeStatusModifier.getStatusIcon(widget.status);
     // this condition is two eliminate the text when we are in XSmall or Small
     return widget.size == OudsBadgeSize.large || widget.size == OudsBadgeSize.medium
         ? SizedBox.expand(
           child: SvgPicture.asset(
             excludeFromSemantics: true,
-            assetName,
+            icon ?? assetName,
             fit: BoxFit.contain,
+            package: icon != null ? OudsTheme.of(context).packageName : null,
             colorFilter: ColorFilter.mode(
-              badgeStatusModifier.getStatusTextAndIconColor((widget.status)),
+              badgeStatusModifier.getStatusTextAndIconColor((widget.status),widget.enabled),
               BlendMode.srcIn,
             ),
           ),